+++ test/unit/vagrant/ssh_test.rb
@@ -197,7 +197,7 @@ class SshTest < Test::Unit::TestCase
     end
 
     should "return false if the connection is refused" do
-      Net::SSH.expects(:start).times(5).raises(Errno::ECONNREFUSED)
+      Net::SSH.expects(:start).times(@env.config.ssh.max_tries).raises(Errno::ECONNREFUSED)
       assert_nothing_raised {
         assert !@ssh.up?
       }